[FLINK-17455][table][filesystem] Move FileSystemFormatFactory to table common
